# openapi.yaml
openapi: 3.1.0
info:
  title: Entities API
  version: 1.2.0
  description: A simple API for managing entities with REST and GraphQL support.
servers:
  - url: http://magdich.cloud:3000
paths:
  /entities:
    get:
      summary: Get all entities
      responses:
        '200':
          description: A list of entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entity'
    post:
      summary: Create a new entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityInput'
      responses:
        '201':
          description: Entity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '400':
          description: Invalid input
  /entities/{id}:
    get:
      summary: Get an entity by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Entity found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '404':
          description: Entity not found
    put:
      summary: Update an entity by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityInput'
      responses:
        '200':
          description: Entity updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '400':
          description: Invalid input
        '404':
          description: Entity not found
    delete:
      summary: Delete an entity by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Entity deleted
        '404':
          description: Entity not found
  /graphql:
    post:
      summary: GraphQL endpoint
      description: GraphQL endpoint for querying and mutating entities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  example: '{ entities { id name size } }'
                variables:
                  type: object
      responses:
        '200':
          description: GraphQL response
          content:
            application/json:
              schema:
                type: object
components:
  schemas:
    Entity:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Sample Entity name
        size:
          type: number
          example: 10
          description: Optional positive number
    EntityInput:
      type: object
      properties:
        name:
          type: string
          example: Sample Entity name
        size:
          type: number
          example: 10
          description: Optional positive number
      required:
        - name